Revert disabling p3a in brave_p3a_service based on feature
The FeatureList singleton isn't yet available at the time
brave_p3a_service::RegisterPrefs is called, so we can't
gate the default value on the feature.

This can be seen by asserting an instance is available:

    DCHECK(base::FeatureList::GetInstance())

Revert the previous attempt, which always disabled P3A regardless
of whether the feature was enabled or not.