Welcome UI: Record P3A opt-in study result on 'done'.
Instead of connecting the opt-in checkbox directly to its own
handler, save it as part of the app state on the js side, and
report it with the other P3A metrics. That way we can postpone
acting on the user choice until the click the 'Done' button at
the end of the workflow as part of explicit confirmation.

Additionally, disable the kP3AEnable pref if it's still in its
default value. The FeatureList isn't available at the time the
pref is registered, so we can't choose the appropriate default
value based on study status.

Since we only disable P3A if the pref is still in its default state,
we won't change the value again after the user has interacted
with it.

TODO: Probably we should set the checkbox to the value of the
pref when we open the page so re-launching the welcome ui doesn't
end up recording a disable.