Interview 'easy': Set Intl.DateTimeFormat().resolvedOptions().timeZone to undefined
Patch v8 to comment out the code that adds the timeZone property to
Intl.DateTimeFormat().resolvedOptions()

Another option would be to delete those rows.

A third altnernative approach could be to write a chromium_src override
for the JSDateTimeFormat::ResolvedOptions that calls the parent
JSDateTimeFormat::ResolvedOptions and then deletes the timeZone property.
This approach has the advantage that it wouldn't require a direct patch.